Update 'Section' component collapsable header to be wholly clickable with no hover state. #141.
