+++ explorer/app/components/Project/components/Contacts/contacts.tsx
@@ -4,12 +4,11 @@ import Link from "next/link";
 import React from "react";
 
 // App dependencies
-import { SectionContent } from "../Section/components/SectionContent/sectionContent";
 import { SectionDetailsEmpty } from "../Section/components/SectionDetailsEmpty/sectionDetailsEmpty";
 import { Section } from "../Section/section";
 
 // Styles
-import { SectionContentListItem } from "../Section/components/SectionContent/sectionContent.styles";
+import { SectionContentListItem } from "../Section/section.styles";
 
 export interface Contact {
   email?: string;
@@ -24,23 +23,21 @@ interface Props {
 export const Contacts = ({ contacts }: Props): JSX.Element => {
   return (
     <Section collapsable title="Contact">
-      <SectionContent gap={4}>
-        {contacts.length > 0 ? (
-          contacts.map(({ email, institution, name }, c) => (
-            <SectionContentListItem key={`${name}${c}`}>
-              <Typography variant="text-body-500-2lines">{name}</Typography>
-              {institution && <span>{institution}</span>}
-              {email && (
-                <Link href={`mailto:${email}`} passHref>
-                  <EmailLink>{email}</EmailLink>
-                </Link>
-              )}
-            </SectionContentListItem>
-          ))
-        ) : (
-          <SectionDetailsEmpty />
-        )}
-      </SectionContent>
+      {contacts.length > 0 ? (
+        contacts.map(({ email, institution, name }, c) => (
+          <SectionContentListItem key={`${name}${c}`}>
+            <Typography variant="text-body-500-2lines">{name}</Typography>
+            {institution && <span>{institution}</span>}
+            {email && (
+              <Link href={`mailto:${email}`} passHref>
+                <EmailLink>{email}</EmailLink>
+              </Link>
+            )}
+          </SectionContentListItem>
+        ))
+      ) : (
+        <SectionDetailsEmpty />
+      )}
     </Section>
   );
 };
