+++ explorer/app/components/Project/components/Description/description.tsx
@@ -3,7 +3,6 @@ import { Typography } from "@mui/material";
 import React from "react";
 
 // App dependencies
-import { SectionContent } from "../Section/components/SectionContent/sectionContent";
 import { Section } from "../Section/section";
 
 interface Props {
@@ -13,9 +12,7 @@ interface Props {
 export const Description = ({ projectDescription }: Props): JSX.Element => {
   return (
     <Section title={"Description"}>
-      <SectionContent gap={4}>
-        <Typography variant="inherit">{projectDescription}</Typography>
-      </SectionContent>
+      <Typography variant="inherit">{projectDescription}</Typography>
     </Section>
   );
 };
