--- explorer/app/components/Project/components/Section/components/SectionContent/sectionContent.styles.ts
@@ -1,17 +1,0 @@
-// Core dependencies
-import styled from "@emotion/styled";
-import { Typography } from "@mui/material";
-
-export const SectionContent = styled(Typography)`
-  align-items: flex-start;
-  color: ${({ theme }) => theme.palette.ink};
-  display: flex;
-  flex-direction: column;
-  padding: 8px 16px 20px;
-`;
-
-export const SectionContentListItem = styled.div`
-  align-items: flex-start;
-  display: flex;
-  flex-direction: column;
-`;
