+++ explorer/app/components/Project/components/Section/section.styles.ts
@@ -1,17 +1,44 @@
-// Core dependencies
 import styled from "@emotion/styled";
-import { ButtonBase } from "@mui/material";
+import { ButtonBase, Typography } from "@mui/material";
+import { BREAKPOINT } from "../../../../hooks/useBreakpointHelper";
 
 export const Section = styled.div`
   background-color: ${({ theme }) => theme.palette.common.white};
   display: flex;
   flex-direction: column;
-  min-height: 64px; /* maintains a minimum height for the collapsed section */
+  padding: 4px 0;
+
+  ${({ theme }) => theme.breakpoints.up(BREAKPOINT.TABLET)} {
+    gap: 8px;
+    padding: 20px;
+  }
 `;
 
 export const SectionSummary = styled(ButtonBase)`
   display: flex;
   justify-content: space-between;
-  margin-bottom: -20px; /* scoots the expanded section content closer to the section summary */
-  padding: 20px 16px; /* maintains a clickable area of 64px for the collapsed section */
+  padding: 16px;
+
+  ${({ theme }) => theme.breakpoints.up(BREAKPOINT.TABLET)} {
+    padding: 0;
+  }
+`;
+
+export const SectionContent = styled(Typography)`
+  align-items: flex-start;
+  color: ${({ theme }) => theme.palette.ink};
+  display: flex;
+  flex-direction: column;
+  gap: 16px;
+  padding: 0 16px 16px;
+
+  ${({ theme }) => theme.breakpoints.up(BREAKPOINT.TABLET)} {
+    padding: 0;
+  }
+`;
+
+export const SectionContentListItem = styled.div`
+  align-items: flex-start;
+  display: flex;
+  flex-direction: column;
 `;
