+++ explorer/app/components/Project/components/Section/section.tsx
@@ -1,7 +1,7 @@
 // Core dependencies
 import RemoveIcon from "@mui/icons-material/Remove";
 import AddIcon from "@mui/icons-material/Add";
-import { ButtonBase, Collapse } from "@mui/material";
+import { Collapse } from "@mui/material";
 import { CollapseProps } from "@mui/material/Collapse/Collapse";
 import React, { ReactNode, useEffect, useState } from "react";
 
@@ -14,7 +14,11 @@ import {
 } from "../../../../hooks/useBreakpointHelper";
 
 // Styles
-import { Section as SectionContainer, SectionSummary } from "./section.styles";
+import {
+  Section as SectionContainer,
+  SectionContent as Content,
+  SectionSummary,
+} from "./section.styles";
 
 interface Props {
   children: ReactNode;
@@ -29,13 +33,16 @@ export const Section = ({
 }: Props): JSX.Element => {
   const mobile = useBreakpointHelper(
     BREAKPOINT_FN_NAME.DOWN,
-    BREAKPOINT.DESKTOP
+    BREAKPOINT.TABLET
   );
   const [expanded, setExpanded] = useState<boolean>(false);
   const [transitionDuration, setTransitionDuration] =
     useState<CollapseProps["timeout"]>(0);
   const disabled = !mobile || !collapsable;
   const ExpandIcon = expanded ? RemoveIcon : AddIcon;
+  const SectionContent = (
+    <Content variant="text-body-400-2lines">{children}</Content>
+  );
 
   const onToggleExpanded = () => {
     setExpanded((expanded) => !expanded);
@@ -48,7 +55,7 @@ export const Section = ({
 
   // Sets collapseTimeout state on change of media breakpoint.
   // Delays setting transitionDuration state for mobile breakpoint to facilitate the immediate transition from
-  //  desktop to mobile.
+  // tablet to mobile.
   useEffect(() => {
     if (mobile) {
       const duration = setTimeout(() => {
@@ -68,12 +75,12 @@ export const Section = ({
         <SectionTitle title={title} />
         {!disabled && <ExpandIcon fontSize="small" />}
       </SectionSummary>
-      {collapsable ? (
+      {!disabled ? (
         <Collapse in={expanded} timeout={transitionDuration}>
-          {children}
+          {SectionContent}
         </Collapse>
       ) : (
-        children
+        SectionContent
       )}
     </SectionContainer>
   );
